+++ src/gromo/modules/linear_growing_module.py
@@ -622,7 +622,7 @@ class LinearGrowingModule(GrowingModule):
     @property
     def tensor_n(self) -> torch.Tensor:
         """
-        Compute the tensor N for the layer with the current M_-2, P and optimal delta.
+        Compute the tensor N for the layer with the current M_{-2}, P and optimal delta.
 
         Returns
         -------
@@ -705,8 +705,8 @@ class LinearGrowingModule(GrowingModule):
             extension of the weight matrix of the layer if None,
             the layer is extended with zeros
             should be of shape:
-             - (out_features, in_features + added_in_features) if added_in_features > 0
-             - (out_features + added_out_features, in_features) if added_out_features > 0
+            - (out_features, in_features + added_in_features) if added_in_features > 0
+            - (out_features + added_out_features, in_features) if added_out_features > 0
         bias_extension: torch.Tensor of shape (out_features + added_out_features,)
             extension of the bias vector of the layer if None,
             the layer is extended with zeros
@@ -1094,90 +1094,3 @@ class LinearGrowingModule(GrowingModule):
                 )
 
         return alpha_weight, alpha_bias, omega, self.eigenvalues_extension
-
-
-if __name__ == "__main__":
-    l1 = LinearGrowingModule(
-        1, 1, use_bias=True, post_layer_function=torch.nn.ReLU(), name="l1"
-    )
-    l2 = LinearGrowingModule(
-        1,
-        1,
-        use_bias=True,
-        previous_module=l1,
-        post_layer_function=torch.nn.ReLU(),
-        name="l2",
-    )
-    l3 = LinearGrowingModule(1, 1, use_bias=True, previous_module=l2, name="l3")
-    x = torch.randn(200, 1, device=global_device())
-    net = torch.nn.Sequential(l1, l2, l3)
-
-    print(net)
-
-    for layer in net:
-        layer.init_computation()
-
-    for layer in net:
-        print(layer.__str__(verbose=1))
-
-    y = net(x)
-    loss = torch.norm(y)
-    print(f"loss: {loss}")
-    loss.backward()
-
-    for layer in net:
-        layer.update_computation()
-
-        layer.compute_optimal_updates()
-
-    for layer in net:
-        layer.reset_computation()
-
-    l1.delete_update()
-    l3.delete_update()
-
-    l2.scaling_factor = 1
-
-    print(f"{l2.first_order_improvement=}")
-    print(f"{l2.weight=}")
-    print(f"{l2.bias=}")
-    print(f"{l2.optimal_delta_layer=}")
-    print(f"{l2.parameter_update_decrease=}")
-    print(f"{l2.extended_input_layer=}")
-    print(f"{l2.extended_input_layer.weight=}")
-    print(f"{l2.extended_input_layer.bias=}")
-    print(f"{l1.extended_output_layer=}")
-    print(f"{l2.eigenvalues_extension=}")
-
-    x_ext = None
-    for layer in net:
-        x, x_ext = layer.extended_forward(x, x_ext)
-
-    new_loss = torch.norm(x)
-    print(f"loss: {new_loss}, {loss - new_loss} improvement")
-    l2.apply_change()
-
-    print("------- New weights -------")
-    print(f"{l1.weight=}")
-    print(f"{l2.weight=}")
-    print(f"{l3.weight=}")
-    print("------- New biases -------")
-    print(f"{l1.bias=}")
-    print(f"{l2.bias=}")
-    print(f"{l3.bias=}")
-
-    for layer in net:
-        layer.init_computation()
-
-    for layer in net:
-        print(layer.__str__(verbose=2))
-
-    y = net(x)
-    loss = torch.norm(y)
-    print(f"loss: {loss}")
-    loss.backward()
-
-    for layer in net:
-        layer.update_computation()
-
-        layer.compute_optimal_updates()